errormanagement

+++ Exercise2/CalcSalary/CalcSalary/Employee.cpp
@@ -39,8 +39,7 @@ void Employee::SetBirthday(Employee::TDate const& birthday) {
 			localtime_s(&ltm, &now);
 
 			//Worker needs to be older than the minimum Age; ltm.tm_years = years since 1900!
-			//cast to prevent overflow!
-			if (((static_cast<int>(ltm.tm_year) + 1900) - minimumAge) >= birthday.year) {
+			if (((ltm.tm_year + 1900) - minimumAge) >= birthday.year) {
 				m_birthday = birthday;
 			}
 			else {
